Refactor bvc into seperate componetns
