+++ src/pages/version/index.ts
@@ -1,13 +1,6 @@
 import { ActionButtons } from "./ActionButtons";
-import { BuildVariants } from "./BuildVariants";
 import { EnqueuePatchModal } from "./EnqueueModal";
 import { Metadata } from "./Metadata";
 import { VersionTabs } from "./VersionTabs";
 
-export {
-  ActionButtons,
-  Metadata,
-  VersionTabs,
-  BuildVariants,
-  EnqueuePatchModal,
-};
+export { ActionButtons, Metadata, VersionTabs, EnqueuePatchModal };
